+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/ProvidesFactoryFirGenerator.kt
@@ -357,9 +357,11 @@ internal class ProvidesFactorySupertypeGenerator(session: FirSession) :
     val callable =
       originClassSymbol.declarationSymbols.filterIsInstance<FirCallableSymbol<*>>().firstOrNull {
         val nameMatches =
-          it.name.asString() == callableName ||
+          it.name.asString().equals(callableName, ignoreCase = true) ||
             (it is FirPropertySymbol &&
-              it.name.asString() == callableName.removePrefix("get").decapitalizeUS())
+              it.name
+                .asString()
+                .equals(callableName.removePrefix("get").decapitalizeUS(), ignoreCase = true))
         if (nameMatches) {
           // Secondary check to ensure it's a @Provides-annotated callable. Otherwise we may
           // match against overloaded non-Provides declarations